requirements.txt
fastapi[all]
uvicorn
redis
nanoid
openai-whisper
